Fix warnings on a 32-bit Debian

+++ far/src/viewer.cpp
@@ -1199,7 +1199,7 @@ int Viewer::ProcessKey(int Key)
 			int64_t NewSelectPos = SelectPos;
 			if (Key == KEY_SHIFTLEFT)
 				--NewSelectPos;
-			fprintf(stderr, "SELECTIO CHANGE: [%ld +%ld)\n", (unsigned long)NewSelectPos, NewSelectSize);
+			fprintf(stderr, "SELECTIO CHANGE: [%lld +%lld)\n", NewSelectPos, NewSelectSize);
 			SelectText(NewSelectPos, NewSelectSize, SelectFlags);
 		}
 		return TRUE;
